Fix elision of literal null values when generating operations from parsed AST
Given graphql that contains literal null values, like the following

```graphql
query(
  $date: DateTime
) {
  events(
    where: {
      occurredAt: $date
      deletedAt: null
    }
  ) {
    id
    eventType
  }
}
```

The alumbra parser correctly parses these null values, but when the
query string is regenerated the null is gone, leading to invalid
graphql. The fix in this commit addresses this in two different
locations. I encountered each of these functions needing to handle
different representations of null values, based on the AST that was
being generated. I am not sure if there are also other representations
in the AST of literal nulls that need to be identified and handled.